requirements.txt
crewai>=0.22.11
langchain>=0.2.0
langchain_community
langchain-groq>=0.1.4
python-dotenv>=1.0.1
streamlit>=1.35.0
